+++ friends/src/components/FriendList.js
@@ -1,98 +1,33 @@
 import React from "react";
-import axios from "axios";
+import { Link } from 'react-router-dom';
+
 import Friend from "./Friend";
-import NewFriendForm from "./NewFriendForm";
 
 export default class FriendList extends React.Component {
-  constructor() {
-    super();
-    this.state = {
-      friends: [],
-      newFriend: {
-        name: "",
-        age: 0,
-        email: ""
-      }
-    };
-  }
-  componentDidMount() {
-    axios
-      .get("http://localhost:5000/friends")
-      .then(response => {
-        this.setState({
-          friends: response.data
-        });
-      })
-      .catch(error => {
-        console.log(error);
-      });
-  }
+    delete = (friend) => {
+      this.props.deleteFriend(friend)
 
-  inputChange = e => {
-    e.preventDefault();
-    this.setState({
-      newFriend: {
-        ...this.state.newFriend,
-        [e.target.name]: e.target.value
-      }
-    });
-  };
-
-  numberInputChange = e => {
-    e.preventDefault();
-    this.setState({
-      newFriend: {
-        ...this.state.newFriend,
-        [e.target.name]: parseInt(e.target.value)
-      }
-    });
-  };
-
-  addNewFriend = (e) => {
-      e.preventDefault();
-    axios
-      .post("http://localhost:5000/friends", this.state.newFriend)
-      .then(response => this.setState({friends: response.data}))
-      .catch(error => console.log(error));
-
-    this.setState({
-      newFriend: {
-        name: "",
-        age: 0,
-        email: ""
-      }
-    });
-  };
-
-    updateFriend = (friend) => {
-    axios.put(`http://localhost:5000/friends/${friend.id}`, friend)
-        .then(response => this.setState({friends: response.data}))
-        .catch(error => console.log(error))
-  };
-
-  deleteFriend = (friend) => {
-      axios.delete(`http://localhost:5000/friends/${friend.id}`)
-        .then(response => this.setState({ friends: response.data}))
-        .catch(error => console.log(error))
   }
 
   render() {
     return (
       <>
-        {this.state.friends.map(friend => (
-          <Friend
-            key={friend.id}
-            friend={friend}
-            updateFriend={this.updateFriend}
-            deleteFriend={this.deleteFriend}
-          />
+        {this.props.friends.map(friend => (
+          <>
+            <div>
+              <div onClick={() => this.delete(friend)}>X</div>
+            </div>
+
+            <Link to={`/friends/${friend.id}`}>
+              <Friend
+                key={friend.id}
+                friend={friend}
+                updateFriend={this.props.updateFriend}
+                deleteFriend={this.props.deleteFriend}
+              />
+            </Link>
+          </>
         ))}
-        <NewFriendForm
-          inputChange={this.inputChange}
-          addNewFriend={this.addNewFriend}
-          numberInputChange={this.numberInputChange}
-          newFriend={this.state.newFriend}
-        />
       </>
     );
   }
